requirements.txt
# LLM Document Processing System - Phase 1-4 Requirements
# Document Processing + Semantic Search + LLM Reasoning + REST API

# ===========================================
# DOCUMENT PROCESSING
# ===========================================
PyMuPDF>=1.23.0        # PDF processing
python-docx>=0.8.11    # DOCX processing  

# ===========================================
# ML/NLP LIBRARIES
# ===========================================
torch>=2.0.0
sentence-transformers>=2.6.1
transformers>=4.40.0
numpy>=1.24.0

# ===========================================
# VECTOR DATABASE
# ===========================================
chromadb>=0.4.0

# ===========================================
# TEXT PROCESSING
# ===========================================
regex>=2023.0.0
nltk>=3.8.1

# ===========================================
# CONFIGURATION & ENVIRONMENT
# ===========================================
pyyaml>=6.0
python-dotenv>=1.0.0

# ===========================================
# UTILITIES
# ===========================================
tqdm>=4.65.0
loguru>=0.7.0

# ===========================================
# TYPE HINTS
# ===========================================
typing-extensions>=4.8.0

# ===========================================
# REST API (Phase 4)
# ===========================================
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9    # file uploads
streamlit>=1.35.0
requests>=2.31.0           # streamlit → API calls
